Renames monitoring entities and modules; Depend on default config root for monitor db instead of tmpfs.
