+++ nucypher/characters/lawful.py
@@ -69,7 +69,7 @@ from nucypher.network.nodes import Teacher
 from nucypher.network.protocols import InterfaceInfo, parse_node_uri
 from nucypher.network.server import ProxyRESTServer, TLSHostingPower, make_rest_app
 from nucypher.blockchain.eth.decorators import validate_checksum_address
-from nucypher.network.status_app.ursula import UrsulaStatusApp
+from nucypher.network.monitor.ursula import UrsulaStatusApp
 
 
 class Alice(Character, BlockchainPolicyAuthor):
