+++ nucypher/network/monitor/assets/favicon.ico
