+++ nucypher/network/monitor/assets/nucypher_logo.png
