+++ nucypher/network/monitor/crawler.py
@@ -1,5 +1,6 @@
 from influxdb import InfluxDBClient
 from maya import MayaDT
+from nucypher.network.status_app.db import BlockchainCrawlerClient
 from twisted.internet import task
 from twisted.logger import Logger
 
@@ -9,8 +10,9 @@ from nucypher.blockchain.eth.agents import ContractAgency, StakingEscrowAgent, N
 from nucypher.blockchain.eth.token import NU, StakeList
 from nucypher.blockchain.eth.utils import datetime_at_period
 from nucypher.config.storages import SQLiteForgetfulNodeStorage
-from nucypher.network.nodes import Learner, FleetStateTracker
-from nucypher.network.status_app.db import BlockchainCrawlerClient
+from nucypher.network.monitor.db import BlockchainCrawlerClient
+from nucypher.network.nodes import FleetStateTracker
+from nucypher.network.nodes import Learner
 
 
 class NetworkCrawler(Learner):
